Modifed view character functions
Okay, so the way I originally did the view characters and the sorting didn't really work. The main issue is that the vectors didn't refer to the actual objects stored in the characters array; so modifications to the actual characters would not show up on the copies stored in the vectors. I could see if there would be some way to correct this, but for the moment I've simply changed the viewParty and viewEnemies functions to work with the characters array. Currently they just show the characters with the appropriate values in the isPartyMem and isEnemy attributes.