Current working version.
Okay, so I have several versions in here already. This is partly because I'm new to gitHub and I'm trying to do local version management too.
Going forward I would like this to be our working version and when proposing changes we can create a new branch for this file.